fixed role check for goals

+++ routes/goals.js
@@ -19,13 +19,23 @@ router.route('/:goalId?')
             }
 
             const email = payload.sub;
-            db.query("SELECT goalId, description, isCompleted FROM mdod.Goal WHERE email = ?;", [email], (error, rows, fields) => {
-                if (error) {
-                    const err = Errors.conflict();
-                    res.status(err.code).json(err);
+
+            //Check if client
+            db.query("SELECT email FROM mdod.Client WHERE email = ?", [email], (err, rows, fields) => {
+                if (rows.length < 1) {
+                    const error = Errors.forbidden();
+                    res.status(error.code).json(error);
                     return;
+                } else {
+                    db.query("SELECT goalId, description, isCompleted FROM mdod.Goal WHERE email = ?;", [email], (error, rows, fields) => {
+                        if (error) {
+                            const err = Errors.conflict();
+                            res.status(err.code).json(err);
+                            return;
+                        }
+                        res.status(200).json(rows);
+                    });
                 }
-                res.status(200).json(rows);
             });
         });
     })
@@ -43,23 +53,31 @@ router.route('/:goalId?')
 
             const goal = new Goal(description);
 
-            if (goal._description) {
-                db.query("INSERT INTO mdod.Goal(email, description) VALUES(?, ?)", [email, goal._description], (error, result) => {
-                    if (error) {
-                        console.log(error);
-                        const err = Errors.conflict();
-                        res.status(err.code).json(err);
-                        return;
-                    }
+            db.query("SELECT email FROM mdod.Client WHERE email = ?", [email], (err, rows, fields) => {
+                if (rows.length < 1) {
+                    const error = Errors.forbidden();
+                    res.status(error.code).json(error);
+                    return;
+                } else {
+                    if (goal._description) {
+                        db.query("INSERT INTO mdod.Goal(email, description) VALUES(?, ?)", [email, goal._description], (error, result) => {
+                            if (error) {
+                                console.log(error);
+                                const err = Errors.conflict();
+                                res.status(err.code).json(err);
+                                return;
+                            }
 
-                    res.status(201).json({
-                        goalId: result.insertId,
-                        message: "Goal aangemaakt"
-                    })
-                })
-            } else {
-                res.status(goal.code).json(goal);
-            }
+                            res.status(201).json({
+                                goalId: result.insertId,
+                                message: "Goal aangemaakt"
+                            })
+                        })
+                    } else {
+                        res.status(goal.code).json(goal);
+                    }
+                }
+            });
         });
     })
     .delete((req, res) => {
@@ -81,25 +99,34 @@ router.route('/:goalId?')
 
             // Get the email of the person who would like to delete the goal.
             const email = payload.sub;
-            db.query("DELETE FROM mdod.Goal WHERE goalId = ? AND email = ?", [goalId, email], (error, result) => {
-                if (error) {
-                    console.log(error);
-                    const err = Errors.conflict();
-                    res.status(err.code).json(err);
-                    return;
-                }
 
-                if (result.affectedRows < 1) {
+            db.query("SELECT email FROM mdod.Client WHERE email = ?", [email], (err, rows, fields) => {
+                if (rows.length < 1) {
                     const error = Errors.forbidden();
                     res.status(error.code).json(error);
                     return;
-                }
+                } else {
+                    db.query("DELETE FROM mdod.Goal WHERE goalId = ? AND email = ?", [goalId, email], (error, result) => {
+                        if (error) {
+                            console.log(error);
+                            const err = Errors.conflict();
+                            res.status(err.code).json(err);
+                            return;
+                        }
 
-                res.status(200).json({
-                    message: "Goal verwijderd."
-                })
-            })
-        })
+                        if (result.affectedRows < 1) {
+                            const error = Errors.forbidden();
+                            res.status(error.code).json(error);
+                            return;
+                        }
+
+                        res.status(200).json({
+                            message: "Goal verwijderd."
+                        })
+                    })
+                }
+            });
+        });
     })
     .put((req, res) => {
         const token = global.stripBearerToken(req.header('Authorization'));
@@ -124,24 +151,32 @@ router.route('/:goalId?')
             // The new goal.
             const goal = new Goal(description);
 
-            db.query("UPDATE mdod.Goal SET description = ? WHERE goalId = ? AND email = ?", [goal._description, goalId, email], (error, result) => {
-                if (error) {
-                    console.log(error);
-                    const err = Errors.conflict();
-                    res.status(err.code).json(err);
-                    return;
-                }
-
-                if (result.affectedRows < 1) {
+            db.query("SELECT email FROM mdod.Client WHERE email = ?", [email], (err, rows, fields) => {
+                if (rows.length < 1) {
                     const error = Errors.forbidden();
                     res.status(error.code).json(error);
                     return;
-                }
+                } else {
+                    db.query("UPDATE mdod.Goal SET description = ? WHERE goalId = ? AND email = ?", [goal._description, goalId, email], (error, result) => {
+                        if (error) {
+                            console.log(error);
+                            const err = Errors.conflict();
+                            res.status(err.code).json(err);
+                            return;
+                        }
 
-                res.status(202).json({
-                    message: "Goal geüpdated."
-                })
-            })
+                        if (result.affectedRows < 1) {
+                            const error = Errors.forbidden();
+                            res.status(error.code).json(error);
+                            return;
+                        }
+
+                        res.status(202).json({
+                            message: "Goal geüpdated."
+                        })
+                    });
+                }
+            });
         });
     });
 
@@ -160,22 +195,30 @@ router.put("/update/status", (req, res) => {
 
         const goalId = req.body.goalId;
 
-        db.query("UPDATE mdod.Goal SET isCompleted = ? where email = ? AND goalId = ?", [isCompleted, email, goalId], (error, result) => {
-            if (error) {
-                const err = Errors.conflict();
-                res.status(err.code).json(err);
+        db.query("SELECT email FROM mdod.Client WHERE email = ?", [email], (err, rows, fields) => {
+            if (rows.length < 1) {
+                const error = Errors.forbidden();
+                res.status(error.code).json(error);
                 return;
-            }
+            } else {
+                db.query("UPDATE mdod.Goal SET isCompleted = ? where email = ? AND goalId = ?", [isCompleted, email, goalId], (error, result) => {
+                    if (error) {
+                        const err = Errors.conflict();
+                        res.status(err.code).json(err);
+                        return;
+                    }
 
-            if (result.affectedRows < 1) {
-                const err = Errors.forbidden();
-                res.status(err.code).json(err);
-                return;
-            }
+                    if (result.affectedRows < 1) {
+                        const err = Errors.forbidden();
+                        res.status(err.code).json(err);
+                        return;
+                    }
 
-            res.status(202).json({
-                message: "Status geüpdate"
-            });
+                    res.status(202).json({
+                        message: "Status geüpdate"
+                    });
+                });
+            }
         });
     });
 });
